add AlertDialog

+++ src/features/home-page/DeleteConfirmation.tsx
@@ -1,14 +1,6 @@
 import { deleteDeck } from "@/actions/actions"
 import { Deck } from "@/types"
-import {
-  Dialog,
-  DialogClose,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog"
-import { Button } from "@/components/ui"
+import { AlertDialog } from "@/components/ui"
 
 export const DeleteConfirmation = ({
   deck,
@@ -16,36 +8,23 @@ export const DeleteConfirmation = ({
 }: {
   deck: Deck
   handleClose: () => void
-}) => {
-  return (
-    <Dialog defaultOpen onOpenChange={handleClose}>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle data-testid="delete-confirmation-title">
-            <span>Are you sure you want to delete </span>
-            <span className="block">
-              <span className="text-destructive-700 text-xl dark:text-destructive-400">
-                {deck.name}
-              </span>
-              <span>?</span>
-            </span>
-          </DialogTitle>
-        </DialogHeader>
-        <DialogFooter className="flex-col-reverse">
-          <Button
-            variant="destructive"
-            onClick={async () => {
-              await deleteDeck(deck.id)
-              handleClose()
-            }}
-          >
-            Yes
-          </Button>
-          <DialogClose asChild>
-            <Button variant="ghost">Cancel</Button>
-          </DialogClose>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  )
-}
+}) => (
+  <AlertDialog
+    title={
+      <span>
+        <span>Are you sure you want to delete </span>
+        <div>
+          <span className="text-destructive-700 text-xl dark:text-destructive-400">
+            {deck.name}
+          </span>
+          <span>?</span>
+        </div>
+      </span>
+    }
+    onConfirm={async () => {
+      await deleteDeck(deck.id)
+      handleClose()
+    }}
+    onCancel={handleClose}
+  />
+)
